영상 삭제, 재 업로드, 내용 수정, BottomSheet 메뉴 추가

+++ app/src/main/java/com/boostcamp/dailyfilm/presentation/uploadfilm/UploadFilmActivity.kt
@@ -15,6 +15,9 @@ import androidx.lifecycle.repeatOnLifecycle
 import com.boostcamp.dailyfilm.R
 import com.boostcamp.dailyfilm.databinding.ActivityUploadFilmBinding
 import com.boostcamp.dailyfilm.presentation.BaseActivity
+import com.boostcamp.dailyfilm.presentation.calendar.CalendarActivity
+import com.boostcamp.dailyfilm.presentation.calendar.DateFragment.Companion.KEY_CALENDAR_INDEX
+import com.boostcamp.dailyfilm.presentation.playfilm.PlayFilmFragment.Companion.KEY_DATE_MODEL
 import com.boostcamp.dailyfilm.presentation.selectvideo.SelectVideoActivity
 import com.boostcamp.dailyfilm.presentation.trimvideo.TrimVideoActivity
 import com.boostcamp.dailyfilm.presentation.util.LottieDialogFragment
@@ -47,6 +50,13 @@ class UploadFilmActivity : BaseActivity<ActivityUploadFilmBinding>(R.layout.acti
                     when (state) {
                         is UiState.Success -> {
                             loadingDialogFragment.hideProgressDialog()
+                            setResult(
+                                RESULT_OK, Intent(
+                                    this@UploadFilmActivity, CalendarActivity::class.java
+                                ).apply {
+                                    putExtra(KEY_CALENDAR_INDEX, viewModel.calendarIndex)
+                                    putExtra(KEY_DATE_MODEL, state.item)
+                                })
                             finish()
                         }
                         is UiState.Loading -> {
@@ -89,12 +99,20 @@ class UploadFilmActivity : BaseActivity<ActivityUploadFilmBinding>(R.layout.acti
             repeatOnLifecycle(Lifecycle.State.STARTED) {
                 viewModel.cancelUploadResult.collect {
                     if (it) {
-                        viewModel.infoItem!!.uri.path?.let { uri -> File(uri).delete() }
-                        startActivity(
-                            Intent(this@UploadFilmActivity, TrimVideoActivity::class.java).apply {
-                                putExtra(SelectVideoActivity.DATE_VIDEO_ITEM, viewModel.beforeItem)
-                            }
-                        )
+                        if (viewModel.beforeItem != null) {
+                            viewModel.infoItem!!.uri.path?.let { uri -> File(uri).delete() }
+                            startActivity(
+                                Intent(
+                                    this@UploadFilmActivity,
+                                    TrimVideoActivity::class.java
+                                ).apply {
+                                    putExtra(
+                                        SelectVideoActivity.DATE_VIDEO_ITEM,
+                                        viewModel.beforeItem
+                                    )
+                                }
+                            )
+                        }
                         finish()
                     }
                 }
